Revisions prompted by documentation of what tests actually do.

+++ bin/checkEOSfiles.pl
@@ -5,17 +5,30 @@ use strict;
 #my $directory = shift;
 #my $SELECTED = shift;
 my $ENV_LD_LIBRARY_PATH = $ENV{LD_LIBRARY_PATH};
+my $OP_CONFIGURE = shift;
+my $HOST_NAME = `hostname | cut -f1 -d.`;  # no domain part`
+#my $HOST_NAME = substr $HOST_NAME_STR, 0, index ".", $HOST_NAME_STR;
+chomp($HOST_NAME);
+
+if ("$OP_CONFIGURE" ne "") {
+   $HOST_NAME = $HOST_NAME.$OP_CONFIGURE;
+}
+my $PREFIX = "TestDir/$HOST_NAME";
 my $SELECTED = "/mnt/scr1/NASA/selected";
-my $CORRUPTED = "/mnt/scr1/NASA/corrupted/MLS"; 
-my $CORREPACKED18 = "/mnt/scr1/NASA/corrupted-repacked18/MLS";
-my $CORREPACKED19 = "/mnt/scr1/NASA/corrupted-repacked19/MLS";
-my $REPACKED18 = "/mnt/scr1/NASA/repacked18/";
-my $REPACKED19 = "/mnt/scr1/NASA/repacked19/";
-my $COPIES = "/mnt/scr1/NASA/copies";
+my $CORRUPTED = "/mnt/scr1/NASA/corrupted/MLS";
+my $CORREPACKED18 = "$PREFIX/corrupted-repacked18/MLS";
+my $CORREPACKED19 = "$PREFIX/corrupted-repacked19/MLS";
+my $REPACKED18 = "$PREFIX/repacked18/";
+my $REPACKED19 = "$PREFIX/repacked19/";
+#my $COPIES = "$PREFIX/copies";
+my $COPYDIR = "$PREFIX/copies";
+#my $CORREPACKED18 = "corrupted-repacked18/MLS";
+#my $CORREPACKED19 = "corrupted-repacked19/MLS";
+#my $REPACKED18 = "repacked18";
+#my $REPACKED19 = "repacked19";
+#my $COPYDIR = "/mnt/scr1/NASA/copies";
 my $OOPS = "";
 my $HDF5_DIR = "/mnt/scr1/pre-release/hdf5/";
-my $HOST_NAME_STR = `hostname`;
-my $HOST_NAME = substr $HOST_NAME_STR, 0, index ".", $HOST_NAME_STR;
 
 my $STR16 = $HDF5_DIR."v16/".$HOST_NAME."-strict";
 my $NSTR16 = $HDF5_DIR."v16/".$HOST_NAME."-nostrict";
@@ -23,16 +36,36 @@ my $STR18 = $HDF5_DIR."v180/".$HOST_NAME."-strict";
 my $NSTR18 = $HDF5_DIR."v180/".$HOST_NAME."-nostrict";
 my $STR19 = $HDF5_DIR."v190/".$HOST_NAME."-strict";
 my $NSTR19 = $HDF5_DIR."v190/".$HOST_NAME."-nostrict";
-my $H5CHECK =  $HDF5_DIR."v180-compat/".$HOST_NAME."/bin/h5check";
+#my $H5CHECK =  $HDF5_DIR."v180-compat/".$HOST_NAME."/bin/h5check";
 my @dirs = ($STR16, $NSTR16, $STR18, $NSTR18, $STR19, $NSTR19);
 my @binaries = ("h5ls", "h5dump -H", "h5stat", "h5copy");
 my $retval = 0;
+
+sub evaluate_result {
+   my $RET = shift;
+   my $result = shift;
+   my $directory = shift;
+   my $testfile = shift;
+   my $dir = shift;
+   my $program = shift;
+      print "Return value was $result\n";
+   if (($RET == 1 && $result == 0) || ($RET == 0 && $result != 0)) {
+      print "Return value was $result\n";
+      print "\nError:  $directory/$testfile does not match expectations with $dir/bin/$program!\n\n";
+      $OOPS = $testfile;
+      #return 1;
+   } else {
+      #print "$testfile is ok with $program.\n";
+      print ".";
+   }
+}
 
 sub check_dir {
    my $dir = shift;
    my $directory = shift;
    my $RET = shift;
    my $result;
+   my $COPIES = "";
    #system("setenv LD_LIBRARY_PATH $dir/lib");
    #Add the current library to the previous LD_LIBRARY_PATH
    $ENV{LD_LIBRARY_PATH} = "$ENV_LD_LIBRARY_PATH:$dir/lib";
@@ -46,60 +79,63 @@ sub check_dir {
       print "checking $testfile\n";
       foreach my $program (@binaries) {
       my $cmd;
+      if ($dir eq "$NSTR18" || $dir eq "$STR18" ) {
+         $COPIES = "$COPYDIR/v180";
+      } elsif ($dir eq "$NSTR19" || $dir eq "$STR19") {
+         $COPIES = "$COPYDIR/v180";
+      }
    #   my $cmd = "/mnt/scr1/lrknox/tools/bin/h5check $_";
    #   my $cmd = "./h5dump -H $_";
          # 1.6 has no h5stat and v1.8 h5stat returns 0 even if it fails to
          # open the file.
-         print "Program to test is $program.\n";
+         print "Program to test is $program with expected return value $RET.\n";
          if ($program eq "h5dump -H"
              || (-e "$dir/bin/$program"
                  &&  ($program ne "h5copy"
-                      || ($dir eq "$STR19"
+                      && ($dir eq "$STR19"
                           || $dir eq "$NSTR19"
-                          || $dir eq "$STR18" 
+                          || $dir eq "$STR18"
                            || $dir eq "$NSTR18")
                           && $RET == 0))) {
 
-            if ($program eq "h5copy") {
-               if (-e "$COPIES/$testfile") {
-                  $cmd = `rm "$COPIES/$testfile"`;     
-               }
-               $cmd = "$dir/bin/$program -i $directory/$testfile -o $COPIES/$testfile -s \"/\" -d \"/COPY\" ";
-            } else {
-               $cmd = "$dir/bin/$program $directory/$testfile";
+         print "Program to test in 1st condition is $program.\n";
+            $cmd = "$dir/bin/$program $directory/$testfile";
+         } elsif ($program eq "h5copy" && -e "$dir/bin/$program") {
+            print "Testing h5copy.\n";
+            if (-e "$COPIES/$testfile") {
+               $cmd = `rm "$COPIES/$testfile"`;
+               my $output = `$cmd`;
+               sleep(10);
             }
-            if ($program eq "h5stat") {
-               open(STDERR,">stderr.txt");
-               $_ = `$cmd`;
-               if(-e "stderr.txt") {
-                  $_ = $_.`cat stderr.txt`;
-               }
-            } else {
-               $_ = `$cmd`;
-            }
-            #my $output = `$cmd`;
-            #print $output;
-            if ($program eq "h5stat" && /Unable to traverse object/) {
-               $result = 1;
-            } else {
-               $result = $?;
-            }
-            print "Return value was $result\n";
-            if (($RET == 1 && $result == 0) || ($RET == 0 && $result != 0)) {
-               print "$_ does not match expectations with $program!\n";
-               $OOPS = $testfile;  
-               #return 1;
-            } else {
-               print "$testfile is ok with $program.\n";
-            }
-        # } else {
-        #    print "Skipping h5stat until it returns non-zero on error.";
+            print "Running h5copy on $directory/$testfile\n";
+            $cmd = "$dir/bin/$program -i $directory/$testfile -o $COPIES/$testfile -s \"/\" -d \"/COPY\" ";
+         } else {
+            print "No $program found in $dir\n";
+            next;
          }
+         if ($program eq "h5stat") {
+            open(STDERR,">stderr.txt");
+            $_ = `$cmd`;
+            if(-e "stderr.txt") {
+               $_ = $_.`cat stderr.txt`;
+            }
+         } else {
+            print "Running $cmd\n";
+            #$_ = `$cmd`;
+         }
+         my $output = `$cmd`;
+         #print $output;
+         if ($program eq "h5stat" && /Unable to traverse object/) {
+            $result = 1;
+         } else {
+            $result = $? >> 2;
+         }
+         evaluate_result ($RET, $result, $directory, $testfile, $dir, $program) ;
       }
    }
    print "Done checking files with $dir.\n\n";
 }
-
+           
 sub countfiles {
    my $directory = shift;
    my $count = 0;
@@ -117,12 +153,12 @@ sub repack_file {
    my $inputdir = shift;
    my $outputdir = shift;
    my $RET = shift;
-   my $result;
+   print "Entered repack_file with params $dir, $inputdir, $outputdir, $RET\n";
    print "Count files.\n";
    my $filecount = countfiles($inputdir);
    my $dayofmonth = `date +%e`;
    my $filenum = $dayofmonth % $filecount;
-   $filecount = 0; 
+   $filecount = 0;
    #Add the current library to the previous LD_LIBRARY_PATH
    $ENV{LD_LIBRARY_PATH} = "$ENV_LD_LIBRARY_PATH:$dir/lib";
    system("echo \$LD_LIBRARY_PATH");
@@ -141,14 +177,7 @@ sub repack_file {
       print "Repacking $testfile with $dir/bin/h5repack\n";
       my $cmd = "$dir/bin/h5repack $inputdir/$testfile $outputdir/$testfile";
       my $output = `$cmd`;
-      $result = $?;
-      print "Return value was $result\n";
-      if (($RET == 1 && $result == 0) || ($RET == 0 && $result != 0)) {
-         print "$testfile does not match expectations with h5repack!\n";
-         $OOPS = $testfile;
-      } else {
-            print "$testfile is ok with h5repack.\n";
-      }
+      evaluate_result ($RET, $?, $inputdir, $testfile, $dir, "h5repack") ;
    }
 }
 
@@ -157,7 +186,6 @@ sub check_diff {
    my $repackeddir = shift;
    my $origdir = shift;
    my $RET = shift;
-   my $result;
    #Add the current library to the previous LD_LIBRARY_PATH
    $ENV{LD_LIBRARY_PATH} = "$ENV_LD_LIBRARY_PATH:$dir/lib";
    system("echo \$LD_LIBRARY_PATH");
@@ -172,15 +200,7 @@ sub check_diff {
       my $cmd = "$dir/bin/h5diff $repackeddir/$testfile $origdir/$testfile";
       my $output = `$cmd`;
    #  print $output;
-      $result = $?;
-      print "Return value was $result\n";
-      if (($RET == 1 && $result == 0) || ($RET == 0 && $result != 0)) {
-         print "$_ does not match expectations with h5diff!\n";
-         $OOPS = $testfile;  
-         #return 1;
-      } else {
-         print "$testfile is ok with h5diff.\n";
-      }
+      evaluate_result ($RET, $?, $repackeddir, $testfile, $dir, "h5diff") ;
    }
 }
 
@@ -203,25 +223,41 @@ sub check_diffs {
 
 sub dump_copies {
    my $dir = shift;
-   my $directory = $COPIES;
+   my $COPIES = "";
+   if ($dir eq "$NSTR18" || $dir eq "$STR18" ) {
+      $COPIES = "$COPYDIR/v180";
+   } elsif ($dir eq "$NSTR19" || $dir eq "$STR19") {
+      $COPIES = "$COPYDIR/v180";
+   } else {
+      return;
+   }
    my $result;
    #system("setenv LD_LIBRARY_PATH $dir/lib");
    #Add the current library to the previous LD_LIBRARY_PATH
    $ENV{LD_LIBRARY_PATH} = "$ENV_LD_LIBRARY_PATH:$dir/lib";
    system("echo \$LD_LIBRARY_PATH");
 
-   #opendir DH, "." or die "Couldn't open the current directory:  $!";
-   opendir DH, $directory or die "Couldn't open the current directory:  $!";
+   print "Dumping copies in $COPIES with $dir/bin/h5dump -H.\n";
+   #opendir DH, "." or die "Couldn't open the current COPIES:  $!";
+   opendir DH, $COPIES or die "Couldn't open the current COPIES:  $!";
    while ($_ = readdir(DH)) {
       next unless ($_ =~ /.*\.h5$/) || ($_ =~ /.*\.he5$/);
       my $testfile = $_;
-      my $cmd = "$dir/bin/h5dump -H $directory/$testfile";
-      my $output = `$cmd`;
+      my $cmd = "$dir/bin/h5dump -H $COPIES/$testfile";
+      open(STDERR,">stderr.txt");
+      $_ = `$cmd`;
       $result = $?;
+      my $output;
+      if(-e "stderr.txt") {
+         $output = $_.`cat stderr.txt`;
+      } else {
+         $output = `$cmd`;
+      }
       print "Return value was $result\n";
+      #print $output;
       if ($result != 0) {
          print "Copied file $testfile cannot be dumped by $dir/bin/h5dump.\n";
-         $OOPS = $testfile;  
+         $OOPS = $testfile;
       } else {
          print "Copied file $testfile successfully dumped by $dir/bin/h5dump.\n";
       }
@@ -229,6 +265,7 @@ sub dump_copies {
 }
 
 sub run_h5check {
+   my $dir = shift;
    my $directory = shift;
    my $RET = shift;
    my $result;
@@ -243,29 +280,30 @@ sub run_h5check {
       my $testfile = $_;
       print "checking $testfile\n";
 
-      my $cmd = "$H5CHECK $directory/$testfile";
+      #my $cmd = "$H5CHECK $directory/$testfile";
+      my $cmd = "$dir/bin/h5check $directory/$testfile";
       my $output = `$cmd`;
-   #  print $output;
-      $result = $?;
-      print "Return value was $result\n";
-      if (($RET == 1 && $result == 0) || ($RET == 0 && $result != 0)) {
-         print "$_ does not match expectations with h5check!\n";
-         $OOPS = $testfile;  
-         #return 1;
-      } else {
-         print "$testfile is ok with h5check.\n";
-      }
+      #$result = $?;
+
+
+      #print "$cmd returned $result"."\nOutput was: $output"."\n";
+      evaluate_result ($RET, $?, $directory, $testfile, $dir, "h5check") ;
    }
 }
 
+print "Removing files from $CORREPACKED18\n";
 my $cmd = `rm $CORREPACKED18/*`;
+print "Removing files from $CORREPACKED19\n";
 $cmd = `rm $CORREPACKED19/*`;
+print "Removing files from $REPACKED18\n";
 $cmd = `rm $REPACKED18/*`;
+print "Removing files from $REPACKED19\n";
 $cmd = `rm $REPACKED19/*`;
-$cmd = `rm $COPIES/*`;
+#print "Removing files from $COPYDIR\n";
+#$cmd = `rm $COPYDIR/*`;
 
 
-print "Check corrupted files with strict checking - should not return 0.\n";
+#print "Check corrupted files with strict checking - should not return 0.\n";
 foreach my $dir (@dirs) {
    $_ = $dir;
    next if (/nostrict/);
@@ -275,8 +313,8 @@ if ($OOPS ne "") {
    print "Corrupted files were not all correctly identified:\n";
    print $OOPS;
    $retval += 1;
+   $OOPS = "";
 }
-
 print "Repack the selected files with strict checking - should return 0.\n";
 foreach my $dir (@dirs) {
    $_ = $dir;
@@ -298,6 +336,7 @@ if ($OOPS ne "") {
    $retval += 1;
 }
 
+# Discontinue repacking files and checking repacked files until h5repack is fixed. LK 1/31/2009
 print "Repack corrupted files with non-strict checking - should return 0.\n";
 foreach my $dir (@dirs) {
    $_ = $dir;
@@ -318,7 +357,8 @@ if ($OOPS ne "") {
    print $OOPS;
    $retval += 1;
 }
-  
+
+
 print "Check selected files with strict and non-strict checking - should return 0 if ok.\n";
 foreach my $dir (@dirs) {
    check_dir ($dir, $SELECTED, 0);
@@ -340,48 +380,64 @@ if ($OOPS ne "") {
    print "Corrupt file found:\n";
    print $OOPS;
    $retval += 1;
-}
-
-run_h5check($CORRUPTED, 1);
-if ($OOPS ne "") {
-   print "h5check failed to find expected errors in known corrupted files.\n";
-   print $OOPS;
-   $retval += 1;
-}
-
-run_h5check($SELECTED, 0);
-run_h5check($REPACKED18, 0);
-run_h5check($REPACKED19, 0);
-run_h5check($CORREPACKED18, 0);
-run_h5check($CORREPACKED19, 0);
-if ($OOPS ne "") {
-   print "h5check found file with unexpected errors.\n";
-   print $OOPS;
-   $retval += 1;
+   $OOPS = "";
 }
 
 foreach my $dir (@dirs) {
+   print "Check all files with $dir/bin/h5check.\n";
+   run_h5check($dir, $CORRUPTED, 1);
+   if ($OOPS ne "") {
+      print "h5check failed to find expected errors in known corrupted files.\n";
+      print $OOPS;
+      $retval += 1;
+      $OOPS = "";
+   }
+
+   run_h5check($dir, $SELECTED, 0);
+   run_h5check($dir, $REPACKED18, 0);
+   run_h5check($dir, $REPACKED19, 0);
+   run_h5check($dir, $CORREPACKED18, 0);
+   run_h5check($dir, $CORREPACKED19, 0);
+   if ($OOPS ne "") {
+      print "h5check found file with unexpected errors.\n";
+      print $OOPS;
+      $retval += 1;
+      $OOPS = "";
+   } else {
+      print "All files ok with $dir/bin/h5check.\n";
+   }
+}
+
+# check_diffs runs h5diff on repacked files
+#disable while h5repack is not working.
+foreach my $dir (@dirs) {
    check_diffs ($dir)
-} 
+}
 if ($OOPS ne "") {
    print "h5diff can't open file or repacked file does no match original file.\n";
    print $OOPS;
    $retval += 1;
+   $OOPS = "";
 }
+
+#dump_copies attempts to dump files copied with h5copy
 foreach my $dir (@dirs) {
    $_ = $dir;
    next if (/nostrict/);
    dump_copies ($dir)
-} 
+}
 if ($OOPS ne "") {
    print "File produced by h5copy can't be dumped with h5dump.\n";
    print $OOPS;
    $retval += 1;
+   $OOPS = "";
 }
+
 if ($retval == 0) {
    print "checkfiles.pl works properly and no problem files found in $SELECTED.";
 } else {
-   print $OOPS;
+#   print $OOPS;
 }
 exit $retval;
+
